Create calculating_paint_for_room
Scenario
The application needs to calculate the total interior surface area to be painted, the
amount of paint needed, and the total cost of the job, including supplies and labour. The
executives remind you that a typical room in a house has four walls, with opposite walls being
equal. Walls can be rectangular or square in shape. Some of rooms, however, might have fewer
than four walls or more than four.

Program Requirements
• The application should use functions to ensure the usability of the code.
• When the user opens the application, the program should show a welcome message.
• Ask the user to input the number of rooms to be painted.
• Be sure to indicate what units are being used for all inputs and outputs.
• Ensure that the application collects information for each room separately, asking for the
number and shape of the surfaces to be painted, as well as features such as windows or
doors.
o Ask the customer if the room shape is square (all four walls to be painted),
rectangular (all four walls to be painted) or custom.
o If the user selects rectangular walls, ask them to enter the width, length and height
of the room.
o If the user selects square walls, ask them to enter just one wall dimension.
o If the user selects custom, ask them for the number and dimensions of each wall to
be painted.
• For all room types, the application should ask the user how many windows and doors the
room contains and their dimensions.
• Program outputs must include:
o For each room: the surface area to be painted, how many gallons of paint are
required, and the total cost of painting that room.
o The overall total surface area to be painted in the house, how many gallons of
paint are required, and the total cost of painting the whole house.
Note: Total cost includes the cost of labour, materials and a profit margin.

Program calculations shall be based on:
o The following constants:
 Paint coverage in square feet per gallon (ft2
/gal) = 350 square feet
 Cost of each gallon of paint ($CDN/gal) = $42
 Labour costs (assume labour cost is included in the cost of paint)
o Area of a wall in a rectangle room = (length or width) * height
o Area of a wall in a square room = length * height
o Wall area to be painted = the area of the wall minus the sum of the area of all the
windows and doors on that wall.